updated gatherData
